refactor: handle synthetix-v3 oracles update logic with react use-query, added 5 retries
