+++ packages/trading-widget/src/core-kit/hooks/trading/index.ts
@@ -10,7 +10,7 @@ export { useExchangeRate } from './use-exchange-rate'
 export { useHandleTrade } from './use-handle-trade'
 export { useMinReceiveText } from './use-min-receive-text'
 export { useMaxSlippagePlaceholder } from './use-max-slippage-placeholder'
-export { useSynthetixV3OraclesUpdate } from './use-synthetix-v3-oracles-update'
+export { useSynthetixV3OraclesUpdate } from './synthetix-v3/use-synthetix-v3-oracles-update'
 export { useDepositProjectedEarnings } from './projected-earnings/use-deposit-projected-earnings'
 export { useProjectedEarningsCore } from './projected-earnings/use-projected-earnings-core'
 export { useIsEasySwapperTrading } from './use-is-easy-swapper-trading'
